Fix agreement signature field (#1)
* Remove NPM_TOKEN env-variable requirement

Previously, the package was private, so using it required having the
NPM_TOKEN environment variable set. It is now public, so this is no
longer a requirement, and the code for it can be removed.

* Improve signature field description in agreement

The customer signature field description was previously "{date} -
{orderID}". It has been changed to the more sensible "{field
description} {customer name} - {date}". The order ID has been moved
underneath the title, and also re-added at the very bottom.

Fixes https://trello.com/c/6iXbhaoq/